\name{extremogram-package}
\alias{extremogram-package}
\alias{extremogram}
\docType{package}
\title{
extremogram
}
\description{
The package estimates the sample univariate, cross and return time extremograms. It can also add empirical confidence bands to each of the extremogram plots via a permutation procedure under the assumption that the data are independent. Finally, the stationary bootstrap allows us to construct credible confidence bands for the extremograms.  

Functions: \enumerate{
\item \code{\link{extremogram1}}
\item \code{\link{extremogram2}}
\item \code{\link{extremogramr}}
\item \code{\link{bootconf1}}
\item \code{\link{bootconf2}}
\item \code{\link{bootconfr}}
\item \code{\link{permfn1}}
\item \code{\link{permfn2}}
\item \code{\link{permfnr}}
}
}
\author{
Nadezda Frolova <nfrolova@ualberta.ca>, 
Ivor Cribben <cribben@ualberta.ca>
}
\references{
\enumerate{
            \item Davis, R. A., Mikosch, T., & Cribben, I. (2012). Towards estimating extremal
            serial dependence via the bootstrapped extremogram. Journal of Econometrics,170(1),
            142-152.
            \item Davis, R. A., Mikosch, T., & Cribben, I. (2011). Estimating extremal
            dependence in univariate and multivariate time series via the extremogram.arXiv preprint arXiv:1107.5592.}
}

